# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Enterprise Device Management Service for POS systems, printers, and payment terminals",
        "title": "Device Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Device Service API Support",
            "email": "support@deviceservice.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8084",
    "basePath": "/api/v1",
    "paths": {
...
        "/devices/{id}": {
            "get": {
                "description": "Get device details and current status by device ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Devices"
                ],
                "summary": "Get device details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Device retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Device"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid device ID",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Device not found",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/operations": {
            "get": {
                "description": "Get list of operations with filtering and pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Operations"
                ],
                "summary": "List operations",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Items per page",
                        "name": "per_page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by device ID",
                        "name": "device_id",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "PRINT",
                            "PAYMENT",
                            "SCAN",
                            "STATUS_CHECK",
                            "OPEN_DRAWER",
                            "DISPLAY_TEXT",
                            "BEEP",
                            "REFUND",
                            "CUT"
                        ],
                        "type": "string",
                        "description": "Filter by operation type",
                        "name": "operation_type",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "PENDING",
                            "PROCESSING",
                            "SUCCESS",
                            "FAILED",
                            "TIMEOUT",
                            "CANCELLED"
                        ],
                        "type": "string",
                        "description": "Filter by status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Start date filter (RFC3339)",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End date filter (RFC3339)",
                        "name": "end_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operations retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "operations": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/definitions/model.DeviceOperation"
                                                    }
                                                },
                                                "pagination": {
                                                    "$ref": "#/definitions/service.PaginationResult"
                                                }
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.ConnectionType": {
            "type": "string",
            "enum": [
                "SERIAL",
                "USB",
                "TCP",
                "BLUETOOTH"
            ],
            "x-enum-varnames": [
                "ConnectionTypeSerial",
                "ConnectionTypeUSB",
                "ConnectionTypeTCP",
                "ConnectionTypeBluetooth"
            ]
        },
        "model.Device": {
            "type": "object",
            "properties": {
                "branch_id": {
                    "type": "string"
                },
                "brand": {
                    "$ref": "#/definitions/model.DeviceBrand"
                },
                "capabilities": {
                    "type": "array",
                    "items": {}
                },
                "connection_config": {
                    "$ref": "#/definitions/model.JSONObject"
                },
                "connection_type": {
                    "$ref": "#/definitions/model.ConnectionType"
                },
                "created_at": {
                    "type": "string"
                },
                "device_id": {
                    "type": "string"
                },
                "device_type": {
                    "$ref": "#/definitions/model.DeviceType"
                },
                "error_info": {
                    "$ref": "#/definitions/model.JSONObject"
                },
                "firmware_version": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_ping": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "performance_metrics": {
                    "$ref": "#/definitions/model.JSONObject"
                },
                "status": {
                    "$ref": "#/definitions/model.DeviceStatus"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "model.DeviceBrand": {
            "type": "string",
            "enum": [
                "EPSON",
                "STAR",
                "INGENICO",
                "PAX",
                "CITIZEN",
                "BIXOLON",
                "VERIFONE",
                "GENERIC"
            ],
            "x-enum-varnames": [
                "BrandEpson",
                "BrandStar",
                "BrandIngenico",
                "BrandPAX",
                "BrandCitizen",
                "BrandBixolon",
                "BrandVerifone",
                "BrandGeneric"
            ]
        },
        "model.DeviceOperation": {
            "type": "object",
            "properties": {
                "completed_at": {
                    "type": "string"
                },
                "correlation_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "device_id": {
                    "type": "string"
                },
                "duration_ms": {
                    "type": "integer"
                },
                "error_message": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "operation_data": {
                    "$ref": "#/definitions/model.JSONObject"
                },
                "operation_type": {
                    "$ref": "#/definitions/model.OperationType"
                },
                "priority": {
                    "$ref": "#/definitions/model.OperationPriority"
                },
                "result": {
                    "$ref": "#/definitions/model.JSONObject"
                },
                "retry_count": {
                    "type": "integer"
                },
                "started_at": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/model.OperationStatus"
                }
            }
        },
        "model.DeviceStatus": {
            "type": "string",
            "enum": [
                "ONLINE",
                "OFFLINE",
                "ERROR",
                "MAINTENANCE",
                "CONNECTING"
            ],
            "x-enum-varnames": [
                "DeviceStatusOnline",
                "DeviceStatusOffline",
                "DeviceStatusError",
                "DeviceStatusMaintenance",
                "DeviceStatusConnecting"
            ]
        },
        "model.DeviceType": {
            "type": "string",
            "enum": [
                "POS",
                "PRINTER",
                "SCANNER",
                "CASH_REGISTER",
                "CASH_DRAWER",
                "DISPLAY"
            ],
            "x-enum-varnames": [
                "DeviceTypePOS",
                "DeviceTypePrinter",
                "DeviceTypeScanner",
                "DeviceTypeCashRegister",
                "DeviceTypeCashDrawer",
                "DeviceTypeDisplay"
            ]
        },
        "model.JSONObject": {
            "type": "object",
            "additionalProperties": true
        },
        "model.OperationPriority": {
            "type": "integer",
            "enum": [
                1,
                2,
                3,
                4,
                5
            ],
            "x-enum-comments": {
                "PriorityBackground": "Bulk operations",
                "PriorityHigh": "Receipt printing, cash drawer",
                "PriorityLow": "Logging, analytics",
                "PriorityNormal": "Status updates, configurations",
                "PriorityUltraCritical": "Payment responses, emergency stops"
            },
            "x-enum-varnames": [
                "PriorityUltraCritical",
                "PriorityHigh",
                "PriorityNormal",
                "PriorityLow",
                "PriorityBackground"
            ]
        },
        "model.OperationStatus": {
            "type": "string",
            "enum": [
                "PENDING",
                "PROCESSING",
                "SUCCESS",
                "FAILED",
                "TIMEOUT",
                "CANCELLED"
            ],
            "x-enum-varnames": [
                "OperationStatusPending",
                "OperationStatusProcessing",
                "OperationStatusSuccess",
                "OperationStatusFailed",
                "OperationStatusTimeout",
                "OperationStatusCancelled"
            ]
        },
        "model.OperationType": {
            "type": "string",
            "enum": [
                "PRINT",
                "PAYMENT",
                "SCAN",
                "STATUS_CHECK",
                "OPEN_DRAWER",
                "DISPLAY_TEXT",
                "BEEP",
                "REFUND",
                "CUT"
            ],
            "x-enum-varnames": [
                "OperationTypePrint",
                "OperationTypePayment",
                "OperationTypeScan",
                "OperationTypeStatusCheck",
                "OperationTypeOpenDrawer",
                "OperationTypeDisplayText",
                "OperationTypeBeep",
                "OperationTypeRefund",
                "OperationTypeCut"
            ]
        },
...
        "service.PaginationResult": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "utils.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "utils.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/utils.APIError"
                },
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}